requirements.txt
google-generativeai
pydantic
tavily-python
rich
matplotlib
typing-extensions
